Add data client, validation
