+++ backend/compact-connect/lambdas/license-data/data_model/schema/__init__.py
@@ -1,3 +1,3 @@
 # We import all the record types with the package to ensure they are all registered
-from .license_schema import LicenseRecordSchema
-from .privilege_schema import PrivilegeRecordSchema
+from .license import LicenseRecordSchema
+from .privilege import PrivilegeRecordSchema
