+++ backend/compact-connect/lambdas/license-data/data_model/schema/license.py
@@ -1,6 +1,6 @@
 from urllib.parse import quote
 
-from marshmallow import pre_dump, post_load
+from marshmallow import pre_dump
 from marshmallow.fields import String, Date
 from marshmallow.validate import Regexp, Length, OneOf
 
@@ -26,7 +26,7 @@ class LicensePublicSchema(LicenseCommonSchema):
     birth_month_day = String(required=False, allow_none=False, validate=Regexp('^[0-1]{1}[0-9]{1}-[0-3]{1}[0-9]{1}'))
 
 
-class LicensePostSchema(LicensePublicSchema):
+class LicensePostSchema(LicenseCommonSchema):
     """
     Schema for license data as posted by a board
     """
@@ -49,12 +49,6 @@ class LicenseRecordSchema(BaseRecordSchema, LicensePostSchema):
     # Generated fields
     upd_ssn = String(required=True, allow_none=False)
     fam_giv_mid_ssn = String(required=True, allow_none=False)
-
-    @post_load
-    def drop_license_gen_fields(self, in_data, **kwargs):  # pylint: disable=unused-argument
-        del in_data['fam_giv_mid_ssn']
-        del in_data['upd_ssn']
-        return in_data
 
     @pre_dump
     def populate_license_gen_fields(self, in_data, **kwargs):  # pylint: disable=unused-argument
