+++ backend/compact-connect/lambdas/license-data/license_csv_reader.py
@@ -2,7 +2,7 @@ from csv import DictReader
 from io import TextIOBase
 from typing import Generator
 
-from data_model.schema.license_schema import LicensePostSchema
+from data_model.schema.license import LicensePostSchema
 
 
 class LicenseCSVReader:
