--- backend/compact-connect/lambdas/license-data/tests/resources/privilege.json
@@ -1,12 +1,0 @@
-{
-  "date_of_issuance": "2024-06-06",
-  "npi": "0608337260",
-  "date_of_birth": "2024-06-06",
-  "license_type": "speech language",
-  "family_name": "string",
-  "middle_name": "string",
-  "status": "active",
-  "ssn": "123-12-1234",
-  "given_name": "string",
-  "home_jurisdiction": "co"
-}
