+++ backend/compact-connect/lambdas/license-data/tests/unit/test_data_model/test_schema/test_license.py
@@ -37,10 +37,4 @@ class TestLicensePostSchema(TstLambdas):
         with open('tests/resources/dynamo/license.json', 'r') as f:
             expected_license_record = json.load(f)
 
-        # These are dynamic and so won't match
-        del expected_license_record['date_of_update']
-        del license_record['date_of_update']
-        del expected_license_record['upd_ssn']
-        del license_record['upd_ssn']
-
         self.assertEqual(expected_license_record, license_record)
